llm-service/requirements.txt
flask>=2.0.0
flask-cors>=3.0.0
torch>=2.0.0
transformers>=4.30.0
scipy>=1.10.0
python-dotenv>=1.0.0
numpy>=1.24.0